+++ NEWS.md
@@ -3,6 +3,7 @@
 * Make connections included in the plot by default.
 * Expose linejoin in ggRelatednessMatrix
 * Expose outline_additive and outline_alpha to ggpedigree
+* Refactored the ggpedigree function to improve readability and maintainability.
 
 # ggpedigree 0.7.1
 * Fixed focal_fill ID being indexed to row order, rather than the ID order.